Re-add support for Rust 1.13.0
