+++ ci/build.sh
@@ -58,7 +58,7 @@ test_target() {
 TARGETS="wasm32-unknown-unknown"
 case "${OS}" in
     linux*)
-        TARGETS="i686-unknown-linux-gnu x86_64-unknown-linux-gnu x86_64-unknown-freebsd x86_64-unknown-netbsd i586-unknown-linux-gnu i586-unknown-linux-musl i686-linux-android i686-unknown-freebsd mipsel-unknown-linux-gnu armv7-linux-androideabi armv7-unknown-linux-gnueabihf armv7-unknown-linux-musleabihf x86_64-sun-solaris "
+        TARGETS="i686-unknown-linux-gnu x86_64-unknown-linux-gnu x86_64-unknown-freebsd x86_64-unknown-netbsd i586-unknown-linux-gnu i586-unknown-linux-musl i686-linux-android i686-unknown-freebsd mipsel-unknown-linux-gnu mips64-unknown-linux-gnuabi64 mips64el-unknown-linux-gnuabi64 mipsel-unknown-linux-gnu mipsel-unknown-linux-musl armv7-linux-androideabi armv7-unknown-linux-gnueabihf armv7-unknown-linux-musleabihf x86_64-sun-solaris aarch64-linux-android powerpc-unknown-linux-gnu powerpc64-unknown-linux-gnu powerpc64le-unknown-linux-gnu s390x-unknown-linux-gnu sparc64-unknown-linux-gnu sparcv9-sun-solaris x86_64-unknown-linux-musl x86_64-linux-android wasm32-unknown-emscripten"
 
         if [ "${RUST}" = "nightly" ]; then
             # aarch64-unknown-cloudabi armv7-unknown-cloudabi-eabihf
