+++ src/redox/mod.rs
@@ -404,3 +404,13 @@ cfg_if! {
         }
     }
 }
+
+cfg_if! {
+    if #[cfg(libc_align)] {
+        mod align;
+        pub use self::align::*;
+    } else {
+        mod no_align;
+        pub use self::no_align::*;
+    }
+}
