+++ src/redox/net.rs
@@ -9,13 +9,6 @@ s! {
         pub s_addr: in_addr_t,
     }
 
-    #[cfg_attr(libc_align, repr(align(4)))]
-    pub struct in6_addr {
-        pub s6_addr: [u8; 16],
-        #[cfg(not(libc_align))]
-        __align: [u32; 0],
-    }
-
     pub struct ip_mreq {
         pub imr_multiaddr: in_addr,
         pub imr_interface: in_addr,
